reorganize reviewer handles getting function
