+++ vibi-dpu/src/core/approval.rs
@@ -1,9 +1,8 @@
 use serde_json::Value;
 
 use crate::core::utils::get_access_token;
+use crate::github::prs::pr_reviewer_handles;
 use crate::{db::review::get_review_from_db, utils::user::ProviderEnum};
-use crate::github::config::prepare_headers;
-use crate::utils::reqwest_client::get_client;
 
 
 
@@ -12,7 +11,8 @@ pub async fn process_approval(deserialised_msg_data: &Value,
     log::debug!("[process_approval] processing approval msg - {:?}", deserialised_msg_data);
     let pr_head_commit = deserialised_msg_data["review"]["commit_id"]
         .to_string().trim_matches('"').to_string();
-    let review_opt = get_review_from_db(&repo_name, &repo_owner, &repo_provider, &pr_number);
+    let review_opt = get_review_from_db(&repo_name,
+        &repo_owner, &repo_provider, &pr_number);
     if review_opt.is_none() {
         log::error!("[process_approval] Unable to get review from db");
         return;
@@ -29,7 +29,8 @@ pub async fn process_approval(deserialised_msg_data: &Value,
     // get reviewer login array by getting pr all reviewer info from gh/bb
     let mut reviewer_handles = Vec::<String>::new();
     if repo_provider == ProviderEnum::Github.to_string().to_lowercase() {
-        let reviewer_handles_opt = get_reviewers_login_handles_for_github_pr(&repo_owner, &repo_name, &pr_number, &pr_head_commit, &final_access_token).await;
+        let reviewer_handles_opt = pr_reviewer_handles(
+            &repo_owner, &repo_name, &pr_number, &pr_head_commit, &final_access_token).await;
         if reviewer_handles_opt.is_none(){
             log::error!("[process_approval] no reviewers handles opt");
             return;
@@ -60,63 +61,4 @@ pub async fn process_approval(deserialised_msg_data: &Value,
     }
     // add up contribution of aliases
     // add comment
-}
-
-pub async fn get_reviewers_login_handles_for_github_pr(repo_owner: &str, repo_name: &str, pr_number: &str, pr_head_commit: &str, access_token: &str) -> Option<Vec<String>> {
-    let headers_opt = prepare_headers(access_token);
-    if headers_opt.is_none() {
-        log::error!("[get_reviewers_login_handles_for_github_pr] Unable to prepare auth headers for repository: {}", repo_name);
-        return None;
-    }
-    let headers = headers_opt.expect("Headers should be present");
-    let client = get_client();
-    let response_result = client
-        .get(&format!(
-            "https://api.github.com/repos/{}/{}/pulls/{}/reviews",
-            repo_owner, repo_name, pr_number
-        ))
-        .headers(headers)
-        .send()
-        .await;
-
-    if response_result.is_err() {
-		let e = response_result.expect_err("No error in sending request");
-		log::error!("[get_reviewers_login_handles_for_github_pr] Failed to send the request: {:?}", e);
-		return None;
-	}
-
-	let response = response_result.expect("Uncaught error in parsing response");
-    if !response.status().is_success() {
-        log::error!(
-            "[get_reviewers_login_handles_for_github_pr] Error in retrieving review list: {:?}",
-            response.status()
-        );
-        return None;
-    }
-
-    let parse_result = response.json::<Vec<Value>>().await;
-    if parse_result.is_err() {
-		let e = parse_result.expect_err("No error in parsing");
-		log::error!(
-			"[get_reviewers_login_handles_for_github_pr] Failed to parse JSON: {:?}",
-			e
-		);
-		return None;
-	}
-	let reviewr_list_result = parse_result.expect("Uncaught error in parsing reviewers list data");
-    // Initialize a vector to store reviewer handles
-    let mut reviewer_handles = Vec::new();
-
-    // Process the review list
-    for review in reviewr_list_result {
-        let state = review["state"].as_str().unwrap_or_default();
-        let commit_id = review["commit_id"].as_str().unwrap_or_default();
-        if state == "APPROVED" && commit_id == pr_head_commit {
-            // Extract reviewer login
-            if let Some(login) = review["user"]["login"].as_str() {
-                reviewer_handles.push(login.to_string());
-            }
-        }
-    }
-    Some(reviewer_handles)
 }
